requirements.txt
line-bot-sdk
pykakasi
requests-oauthlib